Move Unit and ArenaOneByOne with utters classes into separate files
