+++ arenas/one_by_one.py
@@ -1,3 +1,6 @@
+#!/usr/bin/env python3
+# -*- coding: utf-8 -*-
+
 import random
 import time
 
@@ -13,62 +16,17 @@ class Styles(object):
     UNDERLINE = '\033[4m'
 
 
+def report(message):
+    print(Styles.OKBLUE, 'SYSTEM: {}'.format(message), Styles.ENDC)
+
+
 class Dice(object):
-    side = [1, 2, 3, 4, 5, 6]
+    side = range(1, 6)
 
     def throw(self):
         x = random.choice(self.side)
         print('\tDice result:', x)
         return x
-
-
-class Unit(object):
-    # TODO: Add equipment
-    def __init__(self, name, health, damage, dodge, acc):
-        self._name = name
-        self._health = health
-        self._damage = damage
-        self._dodge = dodge
-        self._acc = acc
-
-    @property
-    def name(self):
-        return self._name
-
-    @property
-    def alive(self):
-        return self._health > 0
-
-    @property
-    def health(self):
-        return self._health
-
-    @property
-    def damage(self):
-        return self._damage
-
-    @property
-    def dodge(self):
-        return self._dodge
-
-    @property
-    def acc(self):
-        return self._acc
-
-    @property
-    def status(self):
-        return '{}: health {}, damage {}, dodge {}, acc {}'.format(self.name,
-                                                                   self.health,
-                                                                   self._damage,
-                                                                   self._dodge,
-                                                                   self._acc)
-
-    def hit(self, damage):
-        self._health -= damage
-
-
-def report(message):
-    print(Styles.OKBLUE, 'SYSTEM: {}'.format(message), Styles.ENDC)
 
 
 def attack(unit, attacked, dice_value):
@@ -112,29 +70,3 @@ class ArenaOneByOne(object):
 
         if not self._enemy.alive:
             print(Styles.FAIL, 'WASTED!', Styles.ENDC)
-
-
-def main():
-    hero = Unit(name='Hero',
-                health=10,
-                damage=2,
-                dodge=4,
-                acc=2)
-
-    enemy = Unit(name='Orc',
-                 health=6,
-                 damage=3,
-                 dodge=3,
-                 acc=3)
-
-    arena = ArenaOneByOne(hero, enemy)
-
-    while hero.alive and enemy.alive:
-        arena.move()
-        time.sleep(2)
-
-    print('Fight ended!')
-
-
-if __name__ == '__main__':
-    main()
